Fix image viewer inspector jumpiness
It was caused by the ImageWidget::paintEvent() method attempting to
render the image centered relative to its rect.
The problem, however, is that the Layout occasionally compacts this rect
to a minimum size, after which centering no longer works.
Usually this is done after a resize is complete
or a little while after upddate() is called manually.
The result is that the painter randomly paints centered or to the left,
depending on whether the Layout has compacted its rect or not.
I don't know a good solution for this, but ugly rendering to the left is
preferred over crappy centered rendering.